[adc5g] Removed the need for the high-speed MMCM output on ROACH2.
This fix was originally implemented by Jack Hickish on his fork of mlib_devel
however due to difficulties with merging it's been reimplemented slightly
differently here.

Basically, by using the OVERSAMPLE mode of the ISERDESE1 (instead of the
NETWORKING mode of ISERDES_NODELAY previously used) the need for a high-speed
clock from the MMCM is removed. This makes finding the right MMCM parameters
much easier, in particular at the highest supported speed of ADC5g without
having to do any terrible hacks.

NOTE: the output order when going from NETWORKING->OVERSAMPLE appears
to have changed requiring a swap of the "a" and "d" outputs.