Mode qdr data IDDR onto clk0 domain
** Need to check this is legit ! **

The input QDR data was clocked in with an
IDDR on clk180, and then registered externally into the
clk0 domain, resulting in a clk0_period / 2 timing
constraint. This commit moves the IDDR and sebsequent pipelining
onto clk0, and flips the outputs of the IDDR block.

TODO: test this in hardware.